Crate basic_contract_rs

Main library module

The purpose of this is to do register the contracts.

Other setup can take place here but do so with caution. Any failure will terminate the container and there is limited scope for logging at stage in the lifecyle

Structs

AssetContract
Structure for the AssetContract, on which implemenation transaction functions will be added

MyAsset

Functions

__guest_call

__launch

once_wapc