Struct basic_contract_rs::AssetContract

pub struct AssetContract {}

Structure for the AssetContract, on which implemenation transaction functions will be added

Implementations

impl AssetContract

The contract implementation Should be marked with the macro #[contrant_impl]

All transactions functions should have the #[Transaction(evaluate)] or #[Transaction(submit)] This indicates that the function should be submitted (sent for endorsing to all peers) or is a query operation, that does not need to be written to the ledger.

pub fn new() -> AssetContract

Some form of 'new' function is required. See the lib.rs of this example for how this is used.

pub fn asset_exists(&self, asset_id: String) -> Result<bool, ContractError>

Does the Asset with the supplied key exist?

Returns true or false or ContractError

pub fn invoke_asset_exists(
&self,
args: Vec<WireBuffer>,
return_wb: TypeSchema
) -> Result<WireBuffer, ContractError>

pub fn md_asset_exists(&self) -> TransactionFn

pub fn create_asset(
&self,
my_assset_id: String,
value: String
) -> Result<(), ContractError>

Creates a new asset, with supplied id, and value Marked as submit as this updates the ledger

pub fn invoke_create_asset(
&self,
args: Vec<WireBuffer>,
return_wb: TypeSchema
) -> Result<WireBuffer, ContractError>

pub fn md_create_asset(&self) -> TransactionFn

pub fn read_asset_value(
&self,
my_assset_id: String
) -> Result<String, ContractError>

Reads the value with the supplied asset id

Returns the string value

pub fn invoke_read_asset_value(
&self,
args: Vec<WireBuffer>,
return_wb: TypeSchema
) -> Result<WireBuffer, ContractError>

pub fn md_read_asset_value(&self) -> TransactionFn